Consolidate core database/auth/sync services across clients
