+++ crates/dirt-core/src/lib.rs
@@ -12,6 +12,7 @@ pub mod export;
 pub mod media;
 pub mod models;
 pub mod search;
+pub mod services;
 pub mod state;
 pub mod storage;
 pub mod sync;
