+++ crates/dirt-core/src/services/mod.rs
@@ -1,0 +1,5 @@
+//! Shared service-layer abstractions.
+
+mod database;
+
+pub use database::DatabaseService;
